+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - More unit conversions for external use.
 
+### Fixed
+
+- Correct duplication of z = 0 if at the start or end of the range.
+
 ## 1.0.0 _2022-07-12_
 
 Import first release.